refactor(webfont-generator): separate rendering and output
