+++ packages/webfont-generator/src/rendering/paths.rs
@@ -1,53 +1,4 @@
-use std::io::{Error, ErrorKind};
 use std::path::{Component, Path, PathBuf};
-
-use serde_json::Value;
-
-#[cfg(feature = "napi")]
-use napi::{Error as NapiError, Status};
-
-/// Convert any displayable error into a NAPI GenericFailure error.
-#[cfg(feature = "napi")]
-#[inline]
-pub(crate) fn to_napi_err(error: impl std::fmt::Display) -> NapiError {
-    NapiError::new(Status::GenericFailure, error.to_string())
-}
-
-/// Convert any displayable error into an `io::Error` with `InvalidData` kind.
-#[inline]
-pub(crate) fn to_io_err(error: impl std::fmt::Display) -> Error {
-    Error::new(ErrorKind::InvalidData, error.to_string())
-}
-
-/// Temporarily swap a field in a Handlebars Context, run a render closure,
-/// then restore the original value. Avoids cloning the entire Context.
-#[inline]
-pub(crate) fn render_with_field_swap<F>(
-    ctx: &mut handlebars::Context,
-    key: &str,
-    value: Value,
-    render: F,
-) -> Result<String, Error>
-where
-    F: FnOnce(&handlebars::Context) -> Result<String, Error>,
-{
-    let obj = ctx
-        .data_mut()
-        .as_object_mut()
-        .expect("context should be an object");
-    let original = obj.insert(key.to_owned(), value);
-    let result = render(ctx);
-    let obj = ctx.data_mut().as_object_mut().unwrap();
-    match original {
-        Some(v) => {
-            obj.insert(key.to_owned(), v);
-        }
-        None => {
-            obj.remove(key);
-        }
-    }
-    result
-}
 
 /// Join a base URL with a file name, normalizing slashes.
 ///
